Split up 'model.go' into '{day,event,category}.go'
