+++ src/model/category.go
@@ -1,0 +1,12 @@
+package model
+
+type Category struct {
+	Name     string
+	Priority int
+}
+
+type ByName []Category
+
+func (a ByName) Len() int           { return len(a) }
+func (a ByName) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+func (a ByName) Less(i, j int) bool { return a[i].Name < a[j].Name }
